+++ src/db-types.ts
@@ -30,29 +30,26 @@ interface Tables {
 export type ChangelogRecord = {
   id: number;
   timestamp: number | null;
-}
-
+};
 export type GroupMembersRecord = {
   user_id: number;
   group_id: number;
-}
-
+};
 export type Oauth2ClientsRecord = {
   id: number;
   client_id: string;
   client_secret: string;
   allowed_grant_types: string;
   user_id: number;
-  require_pkce: number | null;
-}
-
+  require_pkce: number;
+};
 export type Oauth2CodesRecord = {
   id: number;
   client_id: number;
   code: string;
   principal_id: number;
-  code_challenge: string | null;
   code_challenge_method: 'plain' | 'S256' | null;
+  code_challenge: string | null;
   created_at: number;
   browser_session_id: string | null;
 
@@ -65,14 +62,12 @@ export type Oauth2CodesRecord = {
    * OpenID Connect Nonce
    */
   nonce: string | null;
-}
-
+};
 export type Oauth2RedirectUrisRecord = {
   id: number;
   oauth2_client_id: number;
   uri: string;
-}
-
+};
 export type Oauth2TokensRecord = {
   id: number;
   oauth2_client_id: number;
@@ -85,16 +80,15 @@ export type Oauth2TokensRecord = {
   browser_session_id: string | null;
 
   /**
-   * 1=implicit, 2=client_credentials, 3=password, 4=authorization_code, 5=authorization_code with secret
+   * 1=implicit, 2=client_credentials, 3=password, 4=authorization_code, 5=authorization_code with secret,6=one-time-token
    */
   grant_type: number | null;
 
   /**
-   * OAuth2 scopes, comma separated
+   * OAuth2 scopes, space separated
    */
   scope: string | null;
-}
-
+};
 export type PrincipalsRecord = {
   id: number;
   identity: string;
@@ -104,30 +98,31 @@ export type PrincipalsRecord = {
   active: number;
 
   /**
-   * 1 = user, 2 = app
+   * 1 = user, 2 = app, 3 = group
    */
   type: number;
   modified_at: number;
-}
 
+  /**
+   * System are built-in and cannot be deleted
+   */
+  system: number;
+};
 export type PrivilegesRecord = {
   privilege: string;
   description: string;
-}
-
+};
 export type ResetPasswordTokenRecord = {
   id: number;
   user_id: number;
   token: string;
   expires_at: number;
   created_at: number;
-}
-
+};
 export type ServerSettingsRecord = {
   setting: string;
   value: string | null;
-}
-
+};
 export type UserAppPermissionsRecord = {
   id: number;
 
@@ -160,8 +155,7 @@ export type UserAppPermissionsRecord = {
    * Last time this application issued or refreshed an access token
    */
   last_used_at: number | null;
-}
-
+};
 export type UserLogRecord = {
   id: number;
   time: number;
@@ -170,27 +164,23 @@ export type UserLogRecord = {
   ip: string;
   user_agent: string | null;
   country: string | null;
-}
-
+};
 export type UserPasswordsRecord = {
   user_id: number;
   password: string;
-}
-
+};
 export type UserPrivilegesRecord = {
   id: number;
   user_id: number;
   resource: string;
   privilege: string;
-}
-
+};
 export type UserTotpRecord = {
   user_id: number;
   secret: string;
   failures: number;
   created: number;
-}
-
+};
 export type UserWebauthnRecord = {
   id: number;
   user_id: number;
@@ -198,8 +188,7 @@ export type UserWebauthnRecord = {
   public_key: string;
   counter: number;
   created: number;
-}
-
+};
 
 
 }